Add Json Web Token, brcryptjs and validator

+++ src/controllers/api.js
@@ -2,6 +2,8 @@ const User = require('../db/models/').User;
 const Budgets = require('../db/models/').Budgets;
 const Category = require('../db/models/').Category;
 const { validationResult } = require('express-validator');
+const jwt = require('jsonwebtoken');
+
 
 let fields =  ['id', 'amount','description', 'type'];
 const getPagination = (page, size) => {
@@ -97,6 +99,26 @@ const api = {
             return res.status(400).json({errors: errors.array()})
         }
         
+        const auth = req.get('authorization');
+
+        let token = "";
+
+        if (auth && auth.toLowerCase().startsWith('bearer')) {
+            token = auth.substring(7)
+        }
+        
+        let decodeToken;
+
+        try {
+            decodeToken = jwt.verify(token, process.env.SECRET)
+        } catch (error) {
+            return res.status(401).json({error: "Invalid token or missing"})
+        }
+
+        if ( !token || !decodeToken.id ) {
+            return res.status(401).json({errors: "Invalid token or missing"})
+        }
+
         Budgets.destroy({
             where: {id: req.params.id}
         })
